Song queue listen endpoint returns updated version of real song queue
