+++ src/server/announcer.py
@@ -1,14 +1,17 @@
 """Defines the anouncer to use for the server sent events and also defines events that can be used consistently."""
-from messageAnnouncer import SSEMessageAnnouncer
+import json
 
-announcer = SSEMessageAnnouncer()
+from .resources import announcer, queue
 
 """Event that sends a whole song queue to a listener."""
 SONG_QUEUE_EVENT = 'song_queue'
 
 def announce_song_queue():
     """Sends the current song queue to all listeners that are listening for the song_queue event as definded above."""
+    song_id_list = []
 
-    """TODO replace with real song queue to send to clients."""
-    song_q = 'Test song queue'
-    announcer.announce(song_q, SONG_QUEUE_EVENT)
+    for votes, song, song_id in queue.get_all():
+        # add song id to list to send out
+        song_id_list.append(song_id)
+
+    announcer.announce(json.dumps(song_id_list), SONG_QUEUE_EVENT)
