+++ test/server/start_server.py
@@ -1,5 +1,5 @@
 """Starts flask server."""
-from main import app
+from src.server.main import app
 
 def start():
     """Start server on localhost, port 5000. For testing purposes."""
