+++ test/server/test_songqueue.py
@@ -2,7 +2,7 @@
 
 import unittest
 
-from songqueue import SongQueue
+from src.server.songqueue import SongQueue
 
 class test_songqueue(unittest.TestCase):
     """Tests for the songqueue functionality."""
