+++ test/server/test_uri_verification.py
@@ -1,7 +1,7 @@
 """Test module for URI verification."""
 import unittest
 
-from validator import validate_uri
+from src.server.validator import validate_uri
 
 class UriVerificationTest(unittest.TestCase):
     """Test Class for all URI verification related tests."""
